requirements.txt
streamlit
pypdf
pdfplumber
fpdf2
reportlab
transformers
spacy
langdetect
deep-translator
pytesseract
Pillow
pandas
matplotlib
gTTS
pyttsx3
pdf2image
PyMuPDF
python-dotenv
torch
